requirements.txt
aiohttp==3.5.0
uvloop==0.11.3
aiohttp-swagger==1.0.5
aiosqlite==0.8.0
cchardet==2.1.4
ujson==1.35
marshmallow==3.0.0rc1
python-json-logger==0.1.10
Pillow==5.3.0
